refactor(ui): Move module declaration to types file
